Add externals to webpack.config for axious
The fix you need is to add axios to the externals option in your webpack config. You don't want webpack bundling axios in your build because axios is already bundled inside js-data-http. If anything, making axios a devDependency (which it is and should be), revealed that you were unnecessarily bundling axios, and this should now allow you to shave a number of kilobytes off your build.